+++ libs/common/RELEASES.md
@@ -1,8 +1,11 @@
-# 0.3.0 
+# 0.3.0
 
 move `ghostos.prompter` to `ghostos_common.prompter`.
 provide prompt object model for `ghosots_moss`.
 
+## 0.3.1
+
+* fix `generate_module_and_attr_name` can not generate from object decorator.
 
 # 0.2.0
 
@@ -11,4 +14,5 @@ provide prompt object model for `ghosots_moss`.
 fix the tree-sitter parser bugs.
 
 ## 0.2.1
+
 prepare uv instead poetry.